requirements.txt
# --- core pins (must-have) ---
numpy==1.24.4
scikit-learn==1.3.2
tqdm
pandas==1.5.3
psycopg2-binary
scipy==1.11.3
reformer_pytorch==1.4.4
stribor==0.1.0
ipython
PyYAML
requests
pyarrow
matplotlib==3.6.0
matplotlib-inline==0.1.6
tensorboard
torchinfo==1.8.0
geotorch==0.3.0

# --- lucidrains compat pins (avoid torch>=2.3 deps) ---
local-attention==1.9.0
product-key-memory==0.1.10
axial-positional-embedding==0.2.1

# NOTE:
# - Do NOT install hyper-connections or colt5-attention.
# - Install torch/torchvision separately and pin them to:
#     torch==2.0.0+cu118
#     torchvision==0.15.1+cu118
#   via: pip install --no-cache-dir "torch==2.0.0+cu118" "torchvision==0.15.1+cu118" \
#        --extra-index-url https://download.pytorch.org/whl/cu118